Remove SPICE Binaries
Removed the SPICE toolkit binaries from the package, instead preferring
to rely on finding them in the system PATH. Added documentation in
installation section requiring installation of SPICE toolkit. I didn't
touch the README because we have a designated ticket for updating it.

Also updated the Dockerfile to install the SPICE toolkit from conda (ARM
or AMD).

Updated tests to search the path rather than hardcoding the bin
directory for SPICE binaries.