Add Minimal de430.bsp Fallback (#38)
This modified de430_earth_sun_moon_20000101_20350101.bsp contains only
the bodies for Earth (399), Earth barycenter (3), Moon (301), and Sun
(10) from 2015-01-01 to 2030-01-01. This file was created with spkmerge
based on the original de430.bsp but is only 1.1MB in size instead of
115MB.

I've also modified the test kernel configs to reference this generic
kernel and they pass because the data for the relevant bodies in the
relevant time ranges is the same.

The idea is that this is a fallback/testing kernel that doesn't blow up
the repo.