Updating instrument intersect ellipsoid
Cleaned up and expanded functionality of the spatial compute function to calculate an intersection with the Earth's ellipsoid.

Expanded the ability to pass in custom pointing vectors without the need to define them in an instrument kernel text format. Specific use case is Libera camera which has ~4 Million pixels and a text based instrument kernel defining them is ~400 MB compared to a ~300 KB netcdf file with the same information.

Naming of functions and parameters along with function structure were updated for clarity and testability.

Testing was expanded and rearranged for clarity. Mocked unit tests were introduced to help test functional usage vs data based results. Integration tests were named as such to use data as an example.